src/model_training.py
import numpy as np
from scipy.stats import uniform, loguniform
from sklearn.model_selection import RandomizedSearchCV
from sklearn.neural_network import MLPRegressor
from utils.logging_utils import setup_logger
from config.config import load_config

def train_model(data):
    logger = setup_logger()
    config = load_config()
    logger.info("Initiating advanced model training with configuration: %s", config)

    X, y = prepare_data(data)
    
    # Neural Architecture Search
    architecture = neural_architecture_search(X, y, config)
    
    # Bayesian Hyperparameter Optimization
    best_params = bayesian_optimization(X, y, architecture, config)
    
    # Final Model Training
    final_model = build_and_train_model(X, y, architecture, best_params)
    
    logger.info("Model training completed successfully")
    return final_model

def prepare_data(data):
    # Implement hyperdimensional encoding here
    X = np.array(data)
    y = np.sin(X)  # Example target function
    return X.reshape(-1, 1), y

def neural_architecture_search(X, y, config):
    logger = setup_logger()
    logger.info("Performing Neural Architecture Search")
    
    search_space = config['neural_architecture_search']['search_space']
    best_architecture = None
    best_score = float('-inf')
    
    for architecture in search_space:
        model = MLPRegressor(hidden_layer_sizes=architecture, max_iter=100)
        model.fit(X, y)
        score = model.score(X, y)
        if score > best_score:
            best_score = score
            best_architecture = architecture
    
    logger.info(f"Best architecture found: {best_architecture}")
    return best_architecture

def bayesian_optimization(X, y, architecture, config):
    logger = setup_logger()
    logger.info("Performing Bayesian Hyperparameter Optimization")
    
    param_distributions = {
        'alpha': loguniform(1e-5, 1e-2),
        'learning_rate_init': loguniform(1e-5, 1e-2),
        'max_iter': uniform(100, 300)
    }
    
    model = MLPRegressor(hidden_layer_sizes=architecture)
    random_search = RandomizedSearchCV(
        model, param_distributions, n_iter=config['bayesian_optimization']['max_iterations'],
        cv=3, scoring='neg_mean_squared_error', n_jobs=-1
    )
    random_search.fit(X, y)
    
    logger.info(f"Best hyperparameters found: {random_search.best_params_}")
    return random_search.best_params_

def build_and_train_model(X, y, architecture, params):
    logger = setup_logger()
    logger.info("Building and training final model")
    
    model = MLPRegressor(hidden_layer_sizes=architecture, **params)
    model.fit(X, y)
    
    return model

if __name__ == "__main__":
    # For testing purposes
    mock_data = np.linspace(0, 10, 100)
    trained_model = train_model(mock_data)
    print("Model training completed. R-squared score:", trained_model.score(mock_data.reshape(-1, 1), np.sin(mock_data)))
