tests/test_model_training.py
import unittest
import numpy as np
from src.model_training import train_model, neural_architecture_search, bayesian_optimization

class TestModelTraining(unittest.TestCase):
    def setUp(self):
        self.mock_data = np.random.rand(100, 10)

    def test_train_model(self):
        trained_model = train_model(self.mock_data)
        self.assertIsNotNone(trained_model)

    def test_neural_architecture_search(self):
        X = self.mock_data
        y = np.random.rand(100)
        config = {'neural_architecture_search': {'search_space': [[64, 32], [128, 64]]}}
        best_architecture = neural_architecture_search(X, y, config)
        self.assertIn(best_architecture, config['neural_architecture_search']['search_space'])

    def test_bayesian_optimization(self):
        X = self.mock_data
        y = np.random.rand(100)
        architecture = [64, 32]
        config = {'bayesian_optimization': {'max_iterations': 10}}
        best_params = bayesian_optimization(X, y, architecture, config)
        self.assertIsInstance(best_params, dict)

if __name__ == '__main__':
    unittest.main()
